datasynth/pyproject.toml
[project]
name = "datasynth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.10.2.post1",
    "pydub>=0.25.1",
    "pytubefix>=7.0.0",
    "pyyaml>=6.0.2",
    "youtube-transcript-api>=0.6.2",
]
